# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "커피 키오스크 API",
    "description": "키오스크·백오피스·마이페이지 API. 인증 필요 엔드포인트는 Authorize에 JWT를 입력하세요.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "로컬 개발" }
  ],
...
  "paths": {
...
    "/api/user/orders": {
      "get": {
        "tags": ["User"],
        "summary": "내 주문 내역",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["ALL", "CANCELED", "WAITING", "COMPLETED", "PENDING"] }, "description": "ALL이 기본" }
        ],
        "responses": {
          "200": { "description": "성공", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/UserOrderItem" } } } } },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/api/admin/orders": {
      "get": {
        "tags": ["Admin"],
        "summary": "주문 목록",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["WAITING", "PREPARING", "PICKUP_READY", "COMPLETED", "CANCELED"] } }
        ],
        "responses": {
          "200": { "description": "성공" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Admin only" }
        }
      }
    },
    "/api/admin/orders/{id}": {
      "patch": {
        "tags": ["Admin"],
        "summary": "주문 상태 변경",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": { "status": { "type": "string", "enum": ["WAITING", "PREPARING", "PICKUP_READY", "COMPLETED", "CANCELED"] } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "성공" },
          "400": { "description": "invalid status" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Admin only" }
        }
      }
    },
...
    "/api/menu": {
      "get": {
        "tags": ["Menu"],
        "summary": "메뉴 목록 (선택: categoryId)",
        "parameters": [{ "name": "categoryId", "in": "query", "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "성공 (MenuItem 형태)" }
        }
      },
      "post": {
        "tags": ["Menu"],
        "summary": "상품 등록 (백오피스)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["categoryId", "name", "basePrice"],
                "properties": {
                  "categoryId": { "type": "string" },
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "basePrice": { "type": "number" },
                  "isSoldOut": { "type": "boolean" },
                  "sortOrder": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "등록 성공" },
          "400": { "description": "categoryId, name, basePrice required" }
        }
      }
    },
    "/api/menu/{id}": {
      "get": {
        "tags": ["Menu"],
        "summary": "메뉴 상세",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "성공" },
          "404": { "description": "Not found" }
        }
      },
      "patch": {
        "tags": ["Menu"],
        "summary": "메뉴 수정",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "basePrice": { "type": "number" },
                  "isSoldOut": { "type": "boolean" },
                  "sortOrder": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "성공" } }
      },
      "delete": {
        "tags": ["Menu"],
        "summary": "메뉴 삭제",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "204": { "description": "삭제 완료" } }
      }
    },
...
    "/api/orders": {
      "post": {
        "tags": ["Orders"],
        "summary": "주문 생성 (키오스크)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["totalPrice", "items"],
                "properties": {
                  "totalPrice": { "type": "number" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["productId", "quantity"],
                      "properties": {
                        "productId": { "type": "string" },
                        "quantity": { "type": "number", "minimum": 1 },
                        "optionIds": { "type": "array", "items": { "type": "string" } }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "주문 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderNumber": { "type": "string" },
                    "orderNo": { "type": "number" },
                    "orderId": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "유효성 오류 (INVALID_TOTAL_PRICE, ITEMS_REQUIRED 등)" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "로그인 응답의 token 값"
      }
    },
    "schemas": {
...
      "UserOrderItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "orderNo": { "type": "number" },
          "orderNumber": { "type": "string" },
          "status": { "type": "string" },
          "totalAmount": { "type": "number" },
          "createdAt": { "type": "string", "format": "date-time" },
          "items": { "type": "array" }
        }
      },
...
    }
  }
}